pyproject.toml
[project]
name = "stockagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "camelot-py[base]>=1.0.0",
    "gradio>=5.46.1",
    "numpy>=1.24.0",
    "opencv-python>=4.12.0.88",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "pytest-playwright>=0.7.1",
    "python-dotenv>=1.1.1",
    "scikit-image>=0.25.2",
    "scrapy>=2.13.3",
    "selenium>=4.34.2",
    "smolagents[litellm]>=1.20.0",
    "torch>=2.8.0",
    "transformers>=4.55.2",
]
